feat(svg-section): imported svg to json and svg code block to json file
